Completed the method that activates the LeaderCards in the PersonalBoard, checking if the Player has the requisite to activate the LeaderCards.
Modified tests that counts the VictoryPoints after completed the method "activateLeaderCard".

Introduced the "amount" attribute to CardRequisite and modified the JsonFile of LeaderCards.

Introduced the Enumeration "RequisiteType" that indicates if the Requisite is a CardRequisite or a ResourceRequisite.

Modified the AddDepot method in the Warehouse that doesn't allow to create to extraDepot of the same Resource.

Added a new method in the Warehouse that counts the total amount of each type of resources.